Install stubs for controlled vocabulary globals during test run
